Ecommerce | Cart | order is lost when more than one items are added within one second (#8968)
* fixed #6464 - Ecommerce | Cart | order is lost when more than one items are added within one second

* Update bundles/CoreBundle/Migrations/Version20210430124911.php

+++ bundles/EcommerceFrameworkBundle/Tools/Installer.php
@@ -68,7 +68,7 @@ class Installer extends AbstractInstaller
               `itemKey` varchar(100) COLLATE utf8_bin NOT NULL,
               `parentItemKey` varchar(100) COLLATE utf8_bin NOT NULL DEFAULT '0',
               `comment` LONGTEXT ASCII,
-              `addedDateTimestamp` int(10) NOT NULL,
+              `addedDateTimestamp` bigint NOT NULL,
               `sortIndex` INT(10) UNSIGNED NULL DEFAULT '0',
               PRIMARY KEY (`itemKey`,`cartId`,`parentItemKey`),
               KEY `cartId_parentItemKey` (`cartId`,`parentItemKey`)
